Move JWT recipe functions into separate interfaces
